requirements.txt
numpy
torch
torchvision
Pillow
matplotlib
selectivesearch